Extracting specific work converters
Moving these to separate files that can be better internalized.
