+++ app/conversions/sipity/conversions/to_rof_hash/specific_work_converters/ulra_converters.rb
@@ -1,0 +1,14 @@
+require 'sipity/conversions/to_rof_hash/specific_work_converters/abstract_converter'
+module Sipity
+  module Conversions
+    module ToRofHash
+      module SpecificWorkConverters
+        class UlraSeniorThesisConverter < AbstractConverter
+        end
+
+        class UlraDocumentConverter < AbstractConverter
+        end
+      end
+    end
+  end
+end
